Refactored to use POJO objects instead of hibernate specific objects. Database specific implementations will be encapsulated in the data access layer.
