+++ zebedee-cms/src/main/java/com/github/onsdigital/zebedee/persistence/dao/impl/DeletedContentFileStore.java
@@ -2,7 +2,7 @@ package com.github.onsdigital.zebedee.persistence.dao.impl;
 
 import com.github.onsdigital.zebedee.exceptions.ZebedeeException;
 import com.github.onsdigital.zebedee.model.ContentWriter;
-import com.github.onsdigital.zebedee.persistence.model.DeletedContentEvent;
+import com.github.onsdigital.zebedee.model.content.deleted.DeletedContentEvent;
 import com.github.onsdigital.zebedee.reader.ContentReader;
 
 import java.io.IOException;
