+++ zebedee-cms/src/main/java/com/github/onsdigital/zebedee/service/DeletedContent/DeletedContentServiceImpl.java
@@ -4,9 +4,9 @@ import com.github.onsdigital.zebedee.content.page.base.Page;
 import com.github.onsdigital.zebedee.exceptions.ZebedeeException;
 import com.github.onsdigital.zebedee.model.Collection;
 import com.github.onsdigital.zebedee.model.ContentWriter;
+import com.github.onsdigital.zebedee.model.content.deleted.DeletedContentEvent;
 import com.github.onsdigital.zebedee.persistence.dao.DeletedContentEventDao;
 import com.github.onsdigital.zebedee.persistence.dao.impl.DeletedContentFileStore;
-import com.github.onsdigital.zebedee.persistence.model.DeletedContentEvent;
 import com.github.onsdigital.zebedee.reader.ContentReader;
 
 import java.io.IOException;
